refactor(Avatar): Move nameToColor in a new helpers file
We need to implement an Avatar like component with Angular for the Cozy Pass extension. We can not reuse the Avatar component from cozy-ui because it is a React component. But we can reuse this nameToColor method to stay as close as we can to cozy-ui.
