+++ react/Avatar/index.jsx
@@ -3,37 +3,11 @@ import cx from 'classnames'
 import PropTypes from 'prop-types'
 import assign from 'lodash/assign'
 
-import palette from '../palette'
 import styles from './styles.styl'
 import Icon, { iconPropType } from '../Icon'
 import PeopleIcon from '../Icons/People'
 import { createSizeStyle } from '../Circle'
-
-export const nameToColor = (name = '') => {
-  const colors = [
-    palette['azure'],
-    palette['melon'],
-    palette['blazeOrange'],
-    palette['pomegranate'],
-    palette['mango'],
-    palette['pumpkinOrange'],
-    palette['darkPeriwinkle'],
-    palette['purpley'],
-    palette['lightishPurple'],
-    palette['weirdGreen'],
-    palette['puertoRico'],
-    palette['emerald'],
-    palette['seafoamGreen'],
-    palette['lavender'],
-    palette['brightSun'],
-    palette['fuchsia']
-  ]
-  const key =
-    Array.from(name.toUpperCase())
-      .map(letter => letter.charCodeAt(0))
-      .reduce((sum, number) => sum + number, 0) % colors.length
-  return colors[key]
-}
+import { nameToColor } from './helpers'
 
 const createBgColorStyle = ({ text, textId }) =>
   text ? { backgroundColor: `${nameToColor(textId || text)}` } : {}
